Changed tables, added registration

+++ demo/demo/src/main/java/com/example/demo/UserRepository.java
@@ -3,6 +3,6 @@ package com.example.demo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByUsername(String username);  // Used for login validation
+public interface UserRepository extends JpaRepository<User, Integer> {  // Changed Long → Integer
+    Optional<User> findByUsername(String username);
 }
